fix: extract coverage-tool dispatch into its own module
The regex literal in extractCoverageTool contains embedded backticks
(matching the `**Coverage toolchain:** `<tool>`` markdown field), which
was confusing Bun's coverage instrumenter's line-mapping for the rest of
check-coverage.ts — the file's own reported line coverage dropped to
~14% despite the code being fully exercised by tests (function-level
coverage stayed at 95%). Moving extractCoverageTool/selectParser (and
their tests) into a dedicated scripts/coverage-tool-dispatch.ts module
resolves the mismapping: check-coverage.ts now reports 99.2% line
coverage and the new module reports 100%, fixing the "coverage must not
decrease" CI gate.
